% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_objects.R
\name{check_effects}
\alias{check_effects}
\title{Check effect size vector}
\usage{
check_effects(es_vector)
}
\arguments{
\item{es_vector}{A vector of user-specified effect sizes.}
}
\value{
If vector has already been added previously the index for
the effect_size object location in the list is return. If it has not, then
the vector is added to the list and 0 returned to indicate that a new
effect_size object should be created.
}
\description{
Checks the vector specifying the effects used for the scenario
to ensure no duplicate effect sizes are being stored.
}
\details{
An internal function that carries out it's operation in the
environment created for effect_size checking. Keeps a list of vectors that
were specified by the user, as well as a list of effect_size objects in this
environment.
}
