% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatingData.R
\name{combine_matrices}
\alias{combine_matrices}
\title{Combines multiple smaller matrices along the diagonal.}
\usage{
combine_matrices(s, m, w, n)
}
\arguments{
\item{s}{A Matrix with a strong correlation structure.}

\item{m}{A Matrix with a moderate correlation structure.}

\item{w}{A Matrix with a weak correlation structure.}

\item{n}{A Matrix with no correlated features. An identity matrix.}
}
\value{
A Matrix that is a combination of all submatrices above.
}
\description{
Only called from create_cor_matrix.
}
