% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatingData.R
\name{create_cor_matrix}
\alias{create_cor_matrix}
\title{Create Custom Correlation Matrix}
\usage{
create_cor_matrix(d, strong = 0, med = 0, weak = 0, none = 0,
  background = 0)
}
\arguments{
\item{d}{A number. The dimensions of the matrix.}

\item{strong}{Number of strong correlations in the matrix. Default: 0}

\item{med}{Number of moderate correlations in the matrix. Default: 0}

\item{weak}{Number of weak correlations in the matrix. Default: 0}

\item{none}{Number of uncorrelated features. Default: 0}

\item{background}{Background correlation amongst all features. Default: 0}
}
\value{
A matrix with the specified correlation structure.
}
\description{
Creates a correlation matrix with a specified number of features
with strong, medium, and weak correlations. Background correlation can be set
by user with background= ; if no specifications given an identity matrix is
returned.
}
\details{
Correlation matrix is built from upper left corner starting with
the first correlation strength specified and working iteratively through
correlation strength specifications along the diagonal.
}
\examples{
create_cor_matrix(5)
create_cor_matrix(10, strong=3, weak=5)
create_core_matrix(20, strong=5, med=7, weak=5, background=0.1)
}
