% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_objects.R
\docType{class}
\name{effect_size-class}
\alias{effect_size-class}
\title{An S4 class to represent the effect size the user specifies is present
between the treatment group and the control group in the scenario.}
\description{
An S4 class to represent the effect size the user specifies is present
between the treatment group and the control group in the scenario.
}
\section{Slots}{

\describe{
\item{\code{effects}}{The vector of differences in mean between the two groups
being considered}

\item{\code{id}}{Numeric assigned by the check_effects method when the object is
created. Unique for each effects vector created.}
}}

