% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_objects.R
\name{effect_size}
\alias{effect_size}
\title{effect_size object constructor}
\usage{
effect_size(effects)
}
\arguments{
\item{effects}{A vector of the effect size the user would like to apply to
the simulation they are running}
}
\value{
An effect size object that can be passed to the simulation functions
}
\description{
Creates (or if it already exists, simply returns) a unique effect size object
}
\details{
Each effect size object is unique. Duplicates will not be created.
}
\examples{
effect_size(c(rep(0,5), 0.4, rep(0.3, 6)))
effect_size(c(0.3, 0.4, 0, 0, 0, 0.2))
}
