% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatingData.R
\name{generate_data}
\alias{generate_data}
\title{Generate data}
\usage{
generate_data(num_samples_trt, num_samples_control, means, delta,
  cor_matrix = NULL, num_reps = 1000)
}
\arguments{
\item{num_samples_trt}{The number of samples in the group that has an effect
applied to the mean specified.}

\item{num_samples_control}{The number of samples in the group that maintains
the mean specified by the user.}

\item{means}{A vector specifying the mean for each feature in the dataset.}

\item{delta}{A vector specifying the difference in the mean for each feature
in the dataset.}

\item{cor_matrix}{The correlation matrix. If not specified an appropriately
sized identity matrix is generated. Default: NULL}

\item{num_reps}{Number of repetitions of the data that will be generated.
Default: 1000}
}
\value{
A dataframe with (num_samples_trt+num_samples_control)*num_reps rows
and length(means) columns as well as a column for sample number, treatment
assignment, iteration number, and total number of samples n.
}
\description{
Generates data using the mvrnorm function from the MASS package
according to user-specified means and effect sizes (ie. differences in means)
for variables.
}
\examples{
generate_data(50, 50, c(1, 2, 3), c(0.3, 0, 0.7))

}
