% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatingData.R
\name{generate_submatrix}
\alias{generate_submatrix}
\title{Generates submatrix with a given correlation structure.}
\usage{
generate_submatrix(dim, lower_bound, upper_bound)
}
\arguments{
\item{dim}{Dimensions of the submatrix}

\item{lower_bound}{Lower boundary for generated random numbers}

\item{upper_bound}{Upper boundary for generated random numbers}
}
\value{
A matrix with a strong, medium, or weak correlation structure.
}
\description{
Called from create_cor_matrix; Uses runif to generate random
numbers between a lower and upper bound (as specified in create_cor_matrix
depending on user's strength indication - strong, med, weak) for the strength
of the correlation.
}
\details{
Designed only to be called by the create_cor_matrix function.
}
