% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{set.alpha}
\alias{set.alpha}
\title{Set alpha for simulation.}
\usage{
set.alpha(alpha)
}
\arguments{
\item{alpha}{The alpha value you wish to use for tests of significance.}
}
\description{
Set alpha for simulation.
}
\examples{
set.alpha(0.05)
}
