% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{single_scenario_simulation}
\alias{single_scenario_simulation}
\title{Simulate single scenario}
\usage{
single_scenario_simulation(num_samples_trt = NULL,
  num_samples_control = NULL, num_genes = NULL, means = NULL,
  delta = NULL, cor_matrix = NULL, num_reps = 1000, method = NULL,
  p_adjust = NULL)
}
\arguments{
\item{num_samples_trt}{A number; specifies how many samples in group exhibiting effects.}

\item{num_samples_control}{A number; specifies how many samples in groups without effects.}

\item{num_genes}{A number; specifies how many features are being analyzed in this scenario.}

\item{means}{A vector which contains the mean for each individual feature.}

\item{delta}{An effect_size object containing the observed effect for each feature.}

\item{cor_matrix}{A matrix which specifies the correlation structure amongst all features.}

\item{num_reps}{A number that describes how many replications of each scenario there will be.}

\item{method}{choose between "ofaat", "mv_glm", "lasso"}

\item{p_adjust}{Used for the internal implementation of p.adjust; takes the same arguments}
}
\value{
A list containing a summary table of overall simulation results,
an individual gene summary result table (with specified effect size and proportion),
and the total number of repetitions that the simulation carried out for this scenario.
}
\description{
Simulates scenario with specified number of samples in treatment and control groups,
specified number of features OR effect size OR mean, a correlation structure between features, method
used for analysis, p value adjustment used in analysis, and the number of repetitions
that should be simulated for this scenario.
}
\section{Notes}{

One of num_samples_trt or num_samples_control must be specified.
One of num_genes, means, or delta must be specified.
Method must be specified.
}

\section{Warning}{

global_alpha must be specified with set.alpha(alpha) for simulation to run to
completion.
}

\examples{
single_scenario_simulation(num_samples_trt = 80, num_samples_control= 95, num_genes= 6, method="ofaat")
}
\seealso{
\code{\link{analyze}} for details on method parameter

\code{\link{p.adjust}} for p value adjustments
}
