% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizingData.R
\name{summarize_overall_power}
\alias{summarize_overall_power}
\title{Summarize overall power}
\usage{
summarize_overall_power(x, truth)
}
\arguments{
\item{x}{A dataframe which contains the summary of selected (1) and not
selected (0) features.}

\item{truth}{A logical vector that states at which positions the user
specified there would be an effect (TRUE), and at which the effect was to be
0 (FALSE).}
}
\value{
A dataframe containing the overall summary results. See ?summary for
more details.
}
\description{
An internal function called by summary.
Processes each row to calculate the false discovery rate in an
accurate way (rather than making the anticipated calculations with the
overall results, the false discoveries for each row are noted and averaged
to get the final result.)
}
\details{
Internal function.
}
