% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizingData.R
\name{summary}
\alias{summary}
\title{Summary}
\usage{
summary(data_to_summarize, e, analysis_method = NULL)
}
\arguments{
\item{data_to_summarize}{Results of analyze function.}

\item{e}{The effect_size object used in data generation. The 'answer key'
to significant features found through analysis.}

\item{analysis_method}{Analysis method used, see analyze documentation for
more information. Default: NULL}
}
\value{
A list of 2. Includes the overall summary dataframe and the
individual feature summary dataframe. Overall summary includes number of
true positives, false positives, true negatives, false negatives, FDR,
TPR/Sensitivity , and where applicable Model_p.value, AUC, and AUC_sd.
Individual feature summary includes the term, proportion of times it was
selected, and the user-specified effect size.
}
\description{
Called after the analyze function within the simulation.
Though analysis is finished this function still requires the method of
analysis used in order to summarize the output correctly.
}
\details{
Best used from within the simulation functions to avoid any
confusion with arguments.
}
\examples{
\dontrun{
if(interactive()){
set.alpha(0.05)
e <- effect_size(c(0.3, 0, 0.7))
example <- generate_data(50, 50, c(1, 2, 3), c(0.3, 0, 0.7))
analyzed_example <- analyze(example, method= "ofaat")
summarized_example <- summary(analyzed_example, e, "ofaat")
}
}
}
\seealso{
\code{\link[dplyr]{select}}
}
