% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtimeTest.R
\name{test_single_scenario_simulation}
\alias{test_single_scenario_simulation}
\title{For run-time testing}
\usage{
test_single_scenario_simulation(num_samples_trt = NULL,
  num_samples_control = NULL, num_genes = NULL, means = NULL,
  delta = NULL, cor_matrix = NULL, num_reps = 1000, method = NULL,
  p_adjust = NULL)
}
\arguments{
\item{num_samples_trt}{Number of samples in group 1. Default: NULL}

\item{num_samples_control}{Number of samples in group 2. Default: NULL}

\item{num_genes}{Number of features being tested. Default: NULL}

\item{means}{A vector of the means for each feature. Default: NULL}

\item{delta}{An effect_size object containing a vector of effects.
Default: NULL}

\item{cor_matrix}{A matrix specifying the correlations. Default: NULL}

\item{num_reps}{Number of iterations of the simulation overall. Default: 1000}

\item{method}{Analysis method used. See ?analyze Default: NULL}

\item{p_adjust}{p value adjustment method used. See ?p.adjust Default: NULL}
}
\value{
Summary of how long each piece of the simulation runs as well as
what kind of analysis method used.
}
\description{
Sys.time() is used to track how long each piece of the
simulation needs to run. Please read ?simulate_single_scenario for detailed
walkthrough of parameters.
}
\details{
For testing purposes only.
}
