% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzingData.R
\name{tidy_glm_mv}
\alias{tidy_glm_mv}
\title{Fit all genes to a general linear model}
\usage{
tidy_glm_mv(p_adjust = NULL, data_tibble)
}
\arguments{
\item{p_adjust}{p value adjustment argument, used to call p.adjust Default: NULL}

\item{data_tibble}{The data to be analyzed}
}
\value{
Returns a dataframe with many columns that are not used in the
summarization step (such as estimate, std.error, statistic, p.value). Main
columns of importance are reps, term, and selected. The "mv_glm" argument
allows a model p.value to be calculated as well.
}
\description{
An internal function. Called by analyze when "mv_glm" argument
is passed.
}
\seealso{
\code{\link[dplyr]{select}}
}
