% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzingData.R
\name{tidy_glm_single}
\alias{tidy_glm_single}
\title{A 'one feature at a time' analysis}
\usage{
tidy_glm_single(p_adjust = NULL, data_tibble)
}
\arguments{
\item{p_adjust}{passed to p.adjust for p value adjustment Default: NULL}

\item{data_tibble}{data to analyze}
}
\value{
Returns a dataframe with many columns that are not used in the
summarization step (such as estimate, std.error, statistic, p.value). Main
columns of importance are reps, term, and selected.
}
\description{
An internal function, called when "ofaat" is the argument
supplied to 'analyze' method.
}
\seealso{
\code{\link[dplyr]{select}}
}
