% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzingData.R
\name{tidy_glmnet}
\alias{tidy_glmnet}
\title{Lasso analysis}
\usage{
tidy_glmnet(data_tibble)
}
\arguments{
\item{data_tibble}{data to be analyzed.}
}
\value{
A 3 column dataframe that tracks the reps number, feature, and
whether that feature was selected as significant or not (either a 1 or 0).
'selected' column also stores the calculated AUC before the summarization
step.
}
\description{
Uses glmnet to perform analysis and calculates AUC.
}
\details{
An internal function.
}
\seealso{
\code{\link[dplyr]{select}}
}
