% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtimeTest.R
\name{test_single_scenario_simulation}
\alias{test_single_scenario_simulation}
\title{For run-time testing}
\usage{
test_single_scenario_simulation(num_samples_trt = NULL,
  num_samples_control = NULL, num_genes = NULL, means = NULL,
  delta = NULL, cor_matrix = NULL, num_reps = 1000, method = NULL,
  p_adjust = NULL)
}
\arguments{
\item{num_samples_trt}{PARAM_DESCRIPTION, Default: NULL}

\item{num_samples_control}{PARAM_DESCRIPTION, Default: NULL}

\item{num_genes}{PARAM_DESCRIPTION, Default: NULL}

\item{means}{PARAM_DESCRIPTION, Default: NULL}

\item{delta}{PARAM_DESCRIPTION, Default: NULL}

\item{cor_matrix}{PARAM_DESCRIPTION, Default: NULL}

\item{num_reps}{PARAM_DESCRIPTION, Default: 1000}

\item{method}{PARAM_DESCRIPTION, Default: NULL}

\item{p_adjust}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
